Enhance PickupController and Blade views for improved shop management

+++ app/Http/Controllers/Admin/PickupController.php
@@ -9,6 +9,8 @@ use App\Models\Shop;
 use Illuminate\Http\Request;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class PickupController extends Controller
 {
@@ -17,9 +19,35 @@ class PickupController extends Controller
      */
     public function index(Request $request): View
     {
-        return view('admin.pickup.index', [
-            'shops' => Shop::whereNot('slug', 'touchvip')->whereNot('slug', 'headquarter')->orderBy('rank', 'asc')->get(),
+        $user = Auth::guard('web')->user();
+
+        $query = Pickup::query();
+        $shop_id = 0;
+        if ($user->hasRole('admin')) {
+            if ($request->has('shop')) {
+                $shop_id = Shop::where('slug', $request->shop)->first()->id;
+                $query->where('shop_id', Shop::where('slug', $request->shop)->first()->id);
+            }else{
+                $shop_id = Shop::where('slug', 'shizuku')->first()->id;
+                $query->where('shop_id', Shop::where('slug', 'shizuku')->first()->id);
+            }
+        }else{
+            $shop_user = DB::connection('mysql')->table('shop_user')->where('user_id', $user->id)->first();
+            $shop_id = $shop_user->shop_id;
+            $query->where('shop_id', $shop_id);
+        }
+        $shops = Shop::whereNot('slug', 'touchvip')->whereNot('slug', 'headquarter')->orderBy('rank', 'asc')->get();
+
+        return view('admin.pickup.detail', [
+            'shop' => Shop::findOrFail($shop_id),
+            'casts' => Cast::where('shop_id', $shop_id)->get(),
+            'pickups' => $query->get(),
+            'shops' => $shops,
         ]);
+
+        // return view('admin.pickup.index', [
+        //     'shops' => Shop::whereNot('slug', 'touchvip')->whereNot('slug', 'headquarter')->orderBy('rank', 'asc')->get(),
+        // ]);
     }
 
     /**
@@ -27,10 +55,18 @@ class PickupController extends Controller
      */
     public function show(Request $request, string $id): View
     {
+        // $query = Pickup::query();
+        // if ($request->has('shop')) {
+        //     $query->where('shop_id', Shop::where('slug', $request->shop)->first()->id);
+        // }else{
+        //     $query->where('shop_id', Shop::where('slug', 'shizuku')->first()->id);
+        // }
+        $shops = Shop::whereNot('slug', 'touchvip')->whereNot('slug', 'headquarter')->orderBy('rank', 'asc')->get();
         return view('admin.pickup.detail', [
             'shop' => Shop::findOrFail($id),
             'casts' => Cast::where('shop_id', $id)->get(),
             'pickups' => Pickup::where('shop_id', $id)->get(),
+            'shops' => $shops,
         ]);
     }
 
@@ -39,9 +75,26 @@ class PickupController extends Controller
      */
     public function update(Request $request, string $id): RedirectResponse
     {
-        Pickup::where('shop_id', $id)->delete();
 
         $pickups = $request->input('pickup', []);
+
+       // 同じcast_idがないかチェック（nullは除外）
+       $nonNullPickups = array_filter($pickups, function($value) {
+        return $value !== null && $value !== '';
+        });
+        $uniquePcikups = array_unique($nonNullPickups);
+        $duplicatePickups = array_diff_assoc($nonNullPickups, $uniquePcikups);
+        if (count($uniquePcikups) !== count($nonNullPickups)) {
+            $duplicateCastIDs = array_unique($duplicatePickups);
+            $duplicateCastNames = array_map(function($value) {
+                return Cast::find($value)->name;
+            }, $duplicateCastIDs);
+            // return redirect()->back()->withInput()->withErrors(['error' => __('message.admin_pickup_error') . 'キャスト名: ' . implode(', ', $duplicateCastNames)]);
+            return redirect()->back()->withInput()->withErrors(['error' => __('message.admin_pickup_error')]);
+        }
+
+        Pickup::where('shop_id', $id)->delete();
+
         foreach ($pickups as $cast_id) {
             if (is_numeric($cast_id)) {
                 Pickup::create([
@@ -51,6 +104,6 @@ class PickupController extends Controller
             }
         }
 
-        return redirect('/admin/pickup/' . $id);
+        return redirect('/admin/pickup/' . $id)->with('success', __('message.admin_pickup_update_success'));
     }
 }
